Fix onHover feature. Separate onHover function. Add getPositionedModuleScopeTypes for getting Hegel typings.
