+++ packages/language-server/server/validation/typings.js
@@ -53,6 +53,7 @@ async function getModuleAST(currentModulePath) {
   );
 }
 
+/** Used in hover feature for desctibing type of node. */
 function getTypeName(type) {
   return type.constraint !== undefined
     ? `${type.name}: ${type.constraint.name}`
